requirements.txt
attrs
dramatiq[redis]
dynaconf
Flask
flask-bootstrap
flask-wtf
flask-sse
freezegun
gevent
gunicorn
hypothesis
pika
psycopg2
pytest
redis
requests
sqlalchemy
tenacity
